feat: Refactor bar chart component
